[Helm] Add the ability to specify container lifecycle (#8010)
**What this PR does / why we need it**:

Add the ability to specify container lifecycle in helm chart. This is
useful when pods are served by a load balancer that needs time to
deregister terminating pods, as we can setup a preStop hook to wait a
bit before terminating the container. Example:

```yaml
gateway:
  lifecycle:
    preStop:
      exec:
        command: ["/bin/sh", "-c", "sleep 30"]
```

**Which issue(s) this PR fixes**:
None

**Special notes for your reviewer**:
None

**Checklist**
- [x] Reviewed the `CONTRIBUTING.md` guide
- [x] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`