Helm enterprise-logs: update kubernetes secrets creation (#8467)
**What this PR does / why we need it**:

Fix the Kubernetes secrets creation for the tokengen and provisioner
jobs for multiple helm installations/uninstallations.

The execution flow for the helm installation is the following:
1) `tokengen` job `loki` container generates a new admin token, stores
it in the object storage and writes it to the `shared` mount.
2) `tokengen` job `create-secret` container creates a new Kubernetes
secret for the generated admin token.
3) `provisioner` job `provisioner` container creates the new admin
resources via the Admin API and writes the tokens to the `bootstrap`
mount.
4) `provisioner` job `create-secret` container creates new Kubernetes
secrets for each token of the additional tenants.

## Previous implementation 

Uninstall and re-install the enterprise-logs helm:
1) `tokengen` job `loki` container generates a new admin token, stores
it in the object storage, and writes it to the `shared` mount. This is
because this job uses an
[emptyDir](https://kubernetes.io/docs/concepts/storage/volumes/#emptydir)
mount which is deleted once the job completes. When a new job is
created, the directory will be again empty, forcing the `tokengen` to
generate a new token.
2) `tokengen` job `create-secret` container fails, because an admin
token secret already exists:
```
Defaulted container "create-secret" out of: create-secret, loki (init)
error: failed to create secret secrets "enterprise-logs-admin-token" already exists
```
3) The helm installation fails ❌ 

## Implementation in this PR

1) `tokengen` job `loki` container generates a new admin token, stores
it in the object storage, and writes it to the `shared` mount.
2) `tokengen` job `create-secret` container updates the previously
created Kubernetes secret for the admin token.
3) `provisioner` job `provisioner` container checks that the admin
resources already exist and doesn't write the token files to the
`bootstrap` mount.
4) `provisioner` job `create-secret` container checks that the token
files are not present on the `bootstrap` mount, and does not create new
secrets.
5) The helm installation succeeds ✅ 

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`