Jsonnet: Fix and update loki-mixin panels (#11529)
**What this PR does / why we need it**:
* Update our legacy panels (Loki/Logs and Loki/Operation) do use
`"type": "timeseries"` instead of `"type": "graph"`. The graph type is
deprecated on recent Grafana versions. I tested the changes with Grizzly
and it worked fine.
* Fix our p99 panel unit. I previously made it to be `s` instead of
`short` but when replicating the change on our source code I put the
change at the wrong identation level. This time I made sure to test the
change with Grizzly.